pyproject.toml
[tool.poetry]
name = "renovels"
version = "0.1.0"
description = ""
authors = ["neytirri <129419810+neytirri@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.23"
greenlet = "^3.0.1"
psycopg = "^3.1.13"
psycopg-binary = "^3.1.13"
pytest = "^7.4.3"
python-multipart = "^0.0.6"
boto3 = "^1.33.11"
gunicorn = "^21.2.0"
argon2-cffi = "^23.1.0"
pydantic = "^2.5.3"
celery = "^5.3.6"
flower = "^2.0.1"
loguru = "^0.7.2"
alembic = "^1.13.1"
fastapi = "^0.108.0"
httpx = "^0.26.0"
uvicorn = "^0.25.0"
pydantic-settings = "^2.1.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
